pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "flatlib>=0.2.3",
    "pydantic>=2.11.7",
    "pyswisseph>=2.8.0.post1",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "timezonefinder>=7.0.2",
    "uvicorn>=0.35.0",
]
